Adding obstacles
It works now... But it's not perfect. The obstacles seem to disappear when contacted, and don't seem to spawn back. So by the end there are barely any in the environment. We should also check they never spawn on top of a robot (though they can probably spawn on top of a goal?).